+++ src/js/de/_3m5/Coco.Init.js
@@ -133,7 +133,7 @@ Coco.SDK = dejavu.Class.declare({
 	////////////////////////////////////////////////////////////
 
 	$statics: {
-		version: "0.1.69",
+		version: "0.1.70",
 		initialized: false
 	},