+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Coco changelog
 
+## 0.1.70
+
+- in progress
+
 ## 0.1.69
 
 - added function removeBy to Coco.Collection